+++ localflavor/generic/checksums.py
@@ -1,10 +1,13 @@
 """Common checksum routines."""
+import warnings
+
 from django.utils import six
+from stdnum import ean as stdnum_ean
+from stdnum import luhn as stdnum_luhn
+
+from localflavor.deprecation import RemovedInLocalflavor30Warning
 
 __all__ = ['luhn', 'ean']
-
-LUHN_ODD_LOOKUP = (0, 2, 4, 6, 8, 1, 3, 5, 7, 9)  # sum_of_digits(index * 2)
-EAN_LOOKUP = (3, 1)
 
 
 def luhn(candidate):
@@ -13,15 +16,25 @@ def luhn(candidate):
 
     Luhn algorithm is used in validation of, for example, credit cards.
     Both numeric and string candidates are accepted.
+
+    .. deprecated:: 2.2
+       Use the luhn function in the python-stdnum_ library instead.
+
+    .. _python-stdnum: https://arthurdejong.org/python-stdnum/
     """
+    warnings.warn(
+        'luhn is deprecated in favor of the luhn function in the python-stdnum library.',
+        RemovedInLocalflavor30Warning,
+    )
+
     if not isinstance(candidate, six.string_types):
         candidate = str(candidate)
-    try:
-        evens = sum(int(c) for c in candidate[-1::-2])
-        odds = sum(LUHN_ODD_LOOKUP[int(c)] for c in candidate[-2::-2])
-        return ((evens + odds) % 10 == 0)
-    except ValueError:  # Raised if an int conversion fails
-        return False
+
+    # Our version returned True for empty strings.
+    if candidate == '':
+        return True
+
+    return stdnum_luhn.is_valid(candidate)
 
 
 def ean(candidate):
@@ -31,16 +44,18 @@ def ean(candidate):
     Note that this validator does not enforce any length checks (usually 13 or 8).
 
     http://en.wikipedia.org/wiki/International_Article_Number_(EAN)
+
+    .. deprecated:: 2.2
+       Use the ean function in the python-stdnum_ library instead.
+
+    .. _python-stdnum: https://arthurdejong.org/python-stdnum/
     """
+    warnings.warn(
+        'ean is deprecated in favor of the ean function in the python-stdnum library.',
+        RemovedInLocalflavor30Warning,
+    )
+
     if not isinstance(candidate, six.string_types):
         candidate = str(candidate)
-    if len(candidate) <= 1:
-        return False
-    given_number, given_checksum = candidate[:-1], candidate[-1]
-    try:
-        calculated_checksum = sum(
-            int(digit) * EAN_LOOKUP[i % 2] for i, digit in enumerate(reversed(given_number)))
-        calculated_checksum = 9 - ((calculated_checksum - 1) % 10)
-        return str(calculated_checksum) == given_checksum
-    except ValueError:  # Raised if an int conversion fails
-        return False
+
+    return stdnum_ean.is_valid(candidate)